fixed the typo that caused the error
The name of the server is mem1 and not _mem1_ as earlier mentioned. I found out the hard way as the code did not run (client didn't connect to server) with the original step. Fixed it now after verifying in Line 56 of NativeFxApp.java program - TextField tf = new TextField("mem1");